// src/frontend/codepeat/src/api-client/models/Reflection.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenBook API
 * Beautiful and Engaging Learning Materials
 *
 * The version of the OpenAPI document: 0.0.1-pre1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Reuse full cleaning and validation logic of the models in the REST API.
 * 
 * Reuse ``full_clean()``, ``clean()``, field validation, and uniqueness checks.
 * Also make sure that the pre-filled model instance can be accessed in the DRF view.
 * @export
 * @interface Reflection
 */
export interface Reflection {
    /**
     * 
     * @type {string}
     * @memberof Reflection
     */
    readonly id: string;
    /**
     * 
     * @type {string}
     * @memberof Reflection
     */
    submission: string;
    /**
     * 
     * @type {any}
     * @memberof Reflection
     */
    answers: any | null;
    /**
     * 
     * @type {Date}
     * @memberof Reflection
     */
    readonly createdAt: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof Reflection
     */
    readonly modifiedAt: Date | null;
}

/**
 * Check if a given object implements the Reflection interface.
 */
export function instanceOfReflection(value: object): value is Reflection {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('submission' in value) || value['submission'] === undefined) return false;
    if (!('answers' in value) || value['answers'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    if (!('modifiedAt' in value) || value['modifiedAt'] === undefined) return false;
    return true;
}

export function ReflectionFromJSON(json: any): Reflection {
    return ReflectionFromJSONTyped(json, false);
}

export function ReflectionFromJSONTyped(json: any, ignoreDiscriminator: boolean): Reflection {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'submission': json['submission'],
        'answers': json['answers'],
        'createdAt': (json['created_at'] == null ? null : new Date(json['created_at'])),
        'modifiedAt': (json['modified_at'] == null ? null : new Date(json['modified_at'])),
    };
}

export function ReflectionToJSON(json: any): Reflection {
    return ReflectionToJSONTyped(json, false);
}

export function ReflectionToJSONTyped(value?: Omit<Reflection, 'id'|'created_at'|'modified_at'> | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'submission': value['submission'],
        'answers': value['answers'],
    };
}