// src/frontend/codepeat/src/api-client/models/ReflectionQuestion.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenBook API
 * Beautiful and Engaging Learning Materials
 *
 * The version of the OpenAPI document: 0.0.1-pre1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { KindEnum } from './KindEnum';
import {
    KindEnumFromJSON,
    KindEnumFromJSONTyped,
    KindEnumToJSON,
    KindEnumToJSONTyped,
} from './KindEnum';

/**
 * Reuse full cleaning and validation logic of the models in the REST API.
 * 
 * Reuse ``full_clean()``, ``clean()``, field validation, and uniqueness checks.
 * Also make sure that the pre-filled model instance can be accessed in the DRF view.
 * @export
 * @interface ReflectionQuestion
 */
export interface ReflectionQuestion {
    /**
     * 
     * @type {string}
     * @memberof ReflectionQuestion
     */
    readonly id: string;
    /**
     * 
     * @type {string}
     * @memberof ReflectionQuestion
     */
    challenge: string;
    /**
     * 
     * @type {string}
     * @memberof ReflectionQuestion
     */
    text: string;
    /**
     * 
     * @type {KindEnum}
     * @memberof ReflectionQuestion
     */
    kind?: KindEnum;
    /**
     * 
     * @type {any}
     * @memberof ReflectionQuestion
     */
    options?: any | null;
    /**
     * 
     * @type {number}
     * @memberof ReflectionQuestion
     */
    position?: number;
    /**
     * 
     * @type {Date}
     * @memberof ReflectionQuestion
     */
    readonly createdAt: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof ReflectionQuestion
     */
    readonly modifiedAt: Date | null;
}



/**
 * Check if a given object implements the ReflectionQuestion interface.
 */
export function instanceOfReflectionQuestion(value: object): value is ReflectionQuestion {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('challenge' in value) || value['challenge'] === undefined) return false;
    if (!('text' in value) || value['text'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    if (!('modifiedAt' in value) || value['modifiedAt'] === undefined) return false;
    return true;
}

export function ReflectionQuestionFromJSON(json: any): ReflectionQuestion {
    return ReflectionQuestionFromJSONTyped(json, false);
}

export function ReflectionQuestionFromJSONTyped(json: any, ignoreDiscriminator: boolean): ReflectionQuestion {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'challenge': json['challenge'],
        'text': json['text'],
        'kind': json['kind'] == null ? undefined : KindEnumFromJSON(json['kind']),
        'options': json['options'] == null ? undefined : json['options'],
        'position': json['position'] == null ? undefined : json['position'],
        'createdAt': (json['created_at'] == null ? null : new Date(json['created_at'])),
        'modifiedAt': (json['modified_at'] == null ? null : new Date(json['modified_at'])),
    };
}

export function ReflectionQuestionToJSON(json: any): ReflectionQuestion {
    return ReflectionQuestionToJSONTyped(json, false);
}

export function ReflectionQuestionToJSONTyped(value?: Omit<ReflectionQuestion, 'id'|'created_at'|'modified_at'> | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'challenge': value['challenge'],
        'text': value['text'],
        'kind': KindEnumToJSON(value['kind']),
        'options': value['options'],
        'position': value['position'],
    };
}